requirements.txt
Flask==2.2.5
ultralytics==8.0.206
pyttsx3==2.90
Pillow==9.5.0
tensorflow==2.11.0
keras==2.11.0
numpy==1.23.5
opencv-python==4.7.0.72
